+++ src/main/kotlin/com/coder/jetbrains/CoderBundle.kt
@@ -1,13 +1,13 @@
-package com.github.codeasher.jetbrainsbackendcoder
+package com.coder.jetbrains
 
 import com.intellij.DynamicBundle
 import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.CoderBundle"
 
-object MyBundle : DynamicBundle(BUNDLE) {
+object CoderBundle : DynamicBundle(BUNDLE) {
 
     @JvmStatic
     fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =